use sealed class
