--- src/main/kotlin/jp/trap/conqest/game/gamestate/PlayingState.kt
@@ -1,7 +1,0 @@
-package jp.trap.conqest.game.gamestate
-
-import jp.trap.conqest.game.GameManager
-import jp.trap.conqest.game.GameState
-
-class PlayingState(private val gameManager: GameManager) : GameState(gameManager) {
-}
